Fix windows build.
TBR=mstarzinger@chromium.org

Review URL: https://codereview.chromium.org/524533002

git-svn-id: https://v8.googlecode.com/svn/branches/bleeding_edge@23508 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ test/cctest/compiler/test-run-properties.cc
@@ -50,7 +50,8 @@ static void TypedArrayLoadHelper(const char* array_type) {
       double value_a = static_cast<U>(kValues[i]);
       double value_b = static_cast<U>(kValues[j]);
       double expected = value_a + value_b;
-      T.CheckCall(T.Val(expected), T.Val(i), T.Val(j));
+      T.CheckCall(T.Val(expected), T.Val(static_cast<double>(i)),
+                  T.Val(static_cast<double>(j)));
     }
   }
 }
